fix(ReactExample): build issues preventing id generation for react-intl
