+++ packages/scene-composer/__mocks__/react-intl.ts
@@ -1,6 +1,19 @@
+const strings = require('../translations/IotAppKitSceneComposer.en_US.json');
+
 const reactIntl = jest.requireActual('react-intl');
-const intl = reactIntl.createIntl({
+type Message = {
+  note: string;
+  text: string;
+};
+
+const messages = Object.entries(strings).reduce((acc, [key, msg]: [key: string, msg: any]) => {
+  acc[key] = msg.text;
+  return acc;
+}, {});
+
+const intl = new reactIntl.createIntl({
   locale: 'en',
+  messages
 });
 
 module.exports = {
