+++ packages/scene-composer/package.json
@@ -40,9 +40,9 @@
     "build:watch": "node ./tools/watch-build.js",
     "dev": "npm-watch build",
     "clean": "rimraf dist ./build/* node_modules",
-    "compile": "run-s svglint extract-msgs convert-svg tttsc copy-assets",
-    "tttsc": "ttsc -p tsconfig.build.json",
-    "tttsc:watch": "ttsc -p tsconfig.build.json --watch",
+    "compile": "run-s svglint extract-msgs convert-svg tsc copy-assets",
+    "tsc": "tsc -p tsconfig.build.json",
+    "tsc:watch": "tsc -p tsconfig.build.json --watch",
     "convert-svg": "npx @svgr/cli --out-dir src/assets/auto-gen/icons/ --typescript --index-template tools/index-template.js -- src/assets/icons/",
     "release": "run-s compile copy-assets",
     "copy-assets": "copyfiles -e \"**/*.tsx\" -e \"**/*.ts\" -e \"**/*.snap\" -e \"**/*.js\" -e \"**/*.jsx\" -e \"**/*.json\" \"src/**/*\" dist/",
@@ -59,7 +59,7 @@
     "start": "npm run storybook",
     "build-storybook": "build-storybook",
     "extract-msgs": "formatjs extract 'src/**/*.ts*' --ignore='**/*.d.ts' --out-file translations/IotAppKitSceneComposer.en_US.json --id-interpolation-pattern '[sha512:contenthash:base64:6]' --format tools/totoro-formatter.js",
-    "prepare": "cd ../.. && husky install",
+    "prepare": "cd ../.. && husky install && ts-patch install -s",
     "copy:license": "cp ../../LICENSE LICENSE",
     "copy:notice": "cp ../../NOTICE NOTICE",
     "prepack": "npm run copy:license && npm run copy:notice",
@@ -129,7 +129,8 @@
     "sass-loader": "^7.3.1",
     "svglint": "^2.2.0",
     "ts-jest": "^27.0.4",
-    "ttypescript": "1.5.13",
+    "ts-patch": "^2.1.0",
+    "ts-transform-react-intl": "^0.4.1",
     "webpack": "^5.76.1"
   },
   "dependencies": {
